Move tracing related files to `tracing` package
