+++ cli/azd/internal/tracing/events/format.go
