+++ cli/azd/internal/tracing/events/format_test.go
