+++ cli/azd/internal/tracing/resource/exec_environment.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/azure/azure-dev/cli/azd/internal"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/fields"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/fields"
 )
 
 func getExecutionEnvironment() string {
