+++ cli/azd/pkg/account/subscriptions_manager.go
@@ -11,9 +11,9 @@ import (
 	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armsubscriptions"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/events"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/fields"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/events"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/fields"
 	"github.com/azure/azure-dev/cli/azd/pkg/auth"
 	"github.com/azure/azure-dev/cli/azd/pkg/convert"
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
@@ -157,7 +157,7 @@ type tenantSubsResult struct {
 // ListSubscription lists subscriptions accessible by the current account by calling azure management services.
 func (m *SubscriptionsManager) ListSubscriptions(ctx context.Context) ([]Subscription, error) {
 	var err error
-	ctx, span := telemetry.GetTracer().Start(ctx, events.AccountSubscriptionsListEvent)
+	ctx, span := tracing.Start(ctx, events.AccountSubscriptionsListEvent)
 	defer span.EndWithStatus(err)
 
 	stop := m.msg.ShowProgress(ctx, "Retrieving subscriptions...")
