+++ cli/azd/pkg/environment/environment.go
@@ -11,8 +11,8 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/fields"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/fields"
 	"github.com/azure/azure-dev/cli/azd/pkg/config"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment/azdcontext"
 	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
@@ -182,11 +182,11 @@ func (e *Environment) Reload() error {
 	}
 
 	if e.GetEnvName() != "" {
-		telemetry.SetUsageAttributes(fields.StringHashed(fields.EnvNameKey, e.GetEnvName()))
+		tracing.SetUsageAttributes(fields.StringHashed(fields.EnvNameKey, e.GetEnvName()))
 	}
 
 	if e.GetSubscriptionId() != "" {
-		telemetry.SetGlobalAttributes(fields.SubscriptionIdKey.String(e.GetSubscriptionId()))
+		tracing.SetGlobalAttributes(fields.SubscriptionIdKey.String(e.GetSubscriptionId()))
 	}
 
 	return nil
@@ -226,7 +226,7 @@ func (e *Environment) Save() error {
 		return fmt.Errorf("saving .env: %w", err)
 	}
 
-	telemetry.SetUsageAttributes(fields.StringHashed(fields.EnvNameKey, e.GetEnvName()))
+	tracing.SetUsageAttributes(fields.StringHashed(fields.EnvNameKey, e.GetEnvName()))
 	return nil
 }
