+++ cli/azd/pkg/project/project.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/azure/azure-dev/cli/azd/internal"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/fields"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/fields"
 	"github.com/azure/azure-dev/cli/azd/pkg/ext"
 	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
 	"github.com/blang/semver/v4"
@@ -111,17 +111,17 @@ func Load(ctx context.Context, projectFilePath string) (*ProjectConfig, error) {
 	if projectConfig.Metadata != nil && projectConfig.Metadata.Template != "" {
 		template := strings.Split(projectConfig.Metadata.Template, "@")
 		if len(template) == 1 { // no version specifier, just the template ID
-			telemetry.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateIdKey, template[0]))
+			tracing.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateIdKey, template[0]))
 		} else if len(template) == 2 { // templateID@version
-			telemetry.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateIdKey, template[0]))
-			telemetry.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateVersionKey, template[1]))
+			tracing.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateIdKey, template[0]))
+			tracing.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateVersionKey, template[1]))
 		} else { // unknown format, just send the whole thing
-			telemetry.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateIdKey, projectConfig.Metadata.Template))
+			tracing.SetUsageAttributes(fields.StringHashed(fields.ProjectTemplateIdKey, projectConfig.Metadata.Template))
 		}
 	}
 
 	if projectConfig.Name != "" {
-		telemetry.SetUsageAttributes(fields.StringHashed(fields.ProjectNameKey, projectConfig.Name))
+		tracing.SetUsageAttributes(fields.StringHashed(fields.ProjectNameKey, projectConfig.Name))
 	}
 
 	if projectConfig.Services != nil {
@@ -137,8 +137,8 @@ func Load(ctx context.Context, projectFilePath string) (*ProjectConfig, error) {
 		slices.Sort(hosts)
 		slices.Sort(languages)
 
-		telemetry.SetUsageAttributes(fields.ProjectServiceLanguagesKey.StringSlice(languages))
-		telemetry.SetUsageAttributes(fields.ProjectServiceHostsKey.StringSlice(hosts))
+		tracing.SetUsageAttributes(fields.ProjectServiceLanguagesKey.StringSlice(languages))
+		tracing.SetUsageAttributes(fields.ProjectServiceHostsKey.StringSlice(hosts))
 	}
 
 	projectConfig.Path = filepath.Dir(projectFilePath)
