+++ cli/azd/pkg/tools/bicep/bicep.go
@@ -15,8 +15,8 @@ import (
 	"runtime"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/events"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/events"
 	"github.com/azure/azure-dev/cli/azd/pkg/config"
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
@@ -182,7 +182,7 @@ func downloadBicep(ctx context.Context, transporter policy.Transporter, bicepVer
 	log.Printf("downloading bicep release %s -> %s", bicepReleaseUrl, name)
 
 	var err error
-	spanCtx, span := telemetry.GetTracer().Start(ctx, events.BicepInstallEvent)
+	spanCtx, span := tracing.Start(ctx, events.BicepInstallEvent)
 	defer span.EndWithStatus(err)
 
 	req, err := http.NewRequestWithContext(spanCtx, "GET", bicepReleaseUrl, nil)
